Move NWConfigParser class to common.py
