+++ nw/__init__.py
@@ -61,8 +61,8 @@ __license__    = "GPLv3"
 __author__     = "Veronica Berglyd Olsen"
 __maintainer__ = "Veronica Berglyd Olsen"
 __email__      = "code@vkbo.net"
-__version__    = "1.5a0"
-__hexversion__ = "0x010500a0"
+__version__    = "1.4rc1"
+__hexversion__ = "0x010400c1"
 __date__       = "2021-07-12"
 __status__     = "Stable"
 __domain__     = "novelwriter.io"
